Make SpaceConstraint::type const
The type of the constraint is bound to the instance. It will never
change for an instantiated object. So we can simply make it const. With
that change the compiler will tell us, when we accidentially try to
change it.

This makes it also necessary to initialize it in the constructor of the
baseclass. This has several benefits. One is: we can enforce to set a
specific type. So we never end up with a unspecified type, as the
compiler will tell us that. Another one is that we can get rid of a
superflous assignment to something "generic" before we set the actual
type.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>